requirements.txt
matplotlib==3.5.3
numpy==1.22.4
pandas==1.4.3
patsy==0.5.2
Pillow==9.3.0
plotly==5.11.0
qrcode==7.3.1
seaborn==0.11.2
setuptools==45.2.0
statsmodels==0.13.2
